python/requirements copy.txt

google-cloud-pubsub
